pyproject.toml
[tool.poetry]
name = "sb_n_eg"
version = "0.1.0"
description = "example of servicebus and eventgrid integration."
authors = ["akitanak <akihito_tanaka@me.com>"]

[tool.poetry.dependencies]
python = ">3.6"
azure-servicebus = "0.50.2"

[tool.poetry.dev-dependencies]
flake8 = "3.7.9"
autopep8 = "1.5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
